Internals Documentation

Index

RauhValladaresEsteban2023.AgeData

RauhValladaresEsteban2023.MicroData

RauhValladaresEsteban2023.ModelParameters

RauhValladaresEsteban2023.aggregatemeans

RauhValladaresEsteban2023.fh′

RauhValladaresEsteban2023.findn

RauhValladaresEsteban2023.fuᴴ

RauhValladaresEsteban2023.fuᵂ

RauhValladaresEsteban2023.gss

RauhValladaresEsteban2023.loadCSVinput

RauhValladaresEsteban2023.loaddeciledata

RauhValladaresEsteban2023.micro2aggregate

RauhValladaresEsteban2023.policyfunction

RauhValladaresEsteban2023.raw2micro

RauhValladaresEsteban2023.simulate

RauhValladaresEsteban2023.simulateindices

RauhValladaresEsteban2023.simulateraw

RauhValladaresEsteban2023.statsf

RauhValladaresEsteban2023.statsf

Internals Interface

RauhValladaresEsteban2023.aggregatemeans — Method

aggregatemeans(v::Vector{MicroData}, w::Vector{Float64})

Computes aggregate means from the vector of MicroData instances v using weights w. Returns an AggregateData instance.

Arguments

v::Vector{MicroData}: A vector of MicroData instances containing agent/individual-level data.

w::Vector{Float64}: A vector of weights. The length of w must be equal to the length of v.

RauhValladaresEsteban2023.fh′ — Method

fh′(h::Float64, a::Float64, n::Float64, p::ModelParameters)::Float64

Returns the value of human capital in the next period.

Arguments

h::Float64: Human capital in the current period.

a::Float64: Ability to acumulate human capital.

n::Float64: Hours worked.

p::ModelParameters: Instance of ModelParameters containing parameter values ϕ and δ.

RauhValladaresEsteban2023.findn — Method

findn(h::Float64, a::Float64, uᴴ::Float64, fV′, p::ModelParameters)

Returns the optimal hours worked and the utility value associated to the optimal hours.

Arguments

h::Float64: Human capital.

a::Float64: Ability to accumulate human capital.

uᴴ::Float64: Utility of staying at home.

fV′: Function computing the continuation value of the next period with respect to hours worked in the current period.

p::ModelParameters: Instance of ModelParameters containing parameter values β.

RauhValladaresEsteban2023.fuᴴ — Method

fuᴴ(h::Float64, κ::Float64, j::Int64, p::ModelParameters)::Float64

Returns the utility value of staying at come.

Arguments

h::Float64: Human capital.

κ::Float64: Realisation of the stay-at-home value shock.

j::Int64: Age.

p::ModelParameters: Instance of ModelParameters containing parameter values ψ, γ, η, κ₀, κ₁, and κ₂.

RauhValladaresEsteban2023.fuᵂ — Method

fuᵂ(h::Float64, n::Float64, p::ModelParameters)::Float64

Returns the utility value of working.

Arguments

h::Float64: Human capital.

n::Float64: Hours worked.

p::ModelParameters: Instance of ModelParameters containing parameter values ω, γ, and ψ.

RauhValladaresEsteban2023.gss — Function

gss(f::Function, a::Float64, b::Float64, tol::Float64 = 1e-6)

Returns argument that maximises value of function f over the interval [a,b] and value of function f at the maximising point using the Golden search method.

Arguments:

f: Univariate function.

a: Lower bound.

b: Upper bound.

tol: Tolerance of the stopping.

RauhValladaresEsteban2023.loadCSVinput — Method

loadCSVinput(inputname::String, group::String)

Loads data from a CSV file located in the model_inputs folder and returns a DataFrame containing data specific to the provided group.

Arguments

inputname::String: The relative path to the CSV file, including the file name. For example, model_inputs/parameters.csv.

group::String: The name of the column in the CSV file that contains the group information. For example: Black.

Notes

The function assumes that the CSV file has a header row containing column names.

The group column in the loaded DataFrame is converted to lowercase for consistent matching.

RauhValladaresEsteban2023.loaddeciledata — Function

loaddeciledata(filename::String, group::String = "")

Loads decile data from a predefined CSV file and returns it as a MicroData structure.

Arguments

filename::String: The path to the CSV file containing the decile data. Example: nlsydata/afqt_decile_1.csv.

group::String: (optional) The group to which the data should be restricted. Valid values are "Black", "White", or an empty string (default).

Returns

A MicroData instance containing the loaded decile data.

Notes

The CSV file is expected to have the following columns: "wage", "hours", "employed", and "weights" (or columns that contain these strings as substrings).

If group is specified, the data will be restricted to the specified group ("Black" or "White") based on the values in the "black" column.

Examples

data = loaddeciledata("nlsydata/afqt_decile_1.csv")  # Load all decile data
data = loaddeciledata("nlsydata/afqt_decile_1.csv", "black")  # Load decile data for the black group only

RauhValladaresEsteban2023.micro2aggregate — Method

micro2aggregate(f, d::MicroData)

Computes aggregate statistics from the MicroData instance d using the specified statistical function f. The function computes aggregate statistics for wage, hours, employment, total income (all individuals), and income (employed individuals).

Arguments

f: A function that computes the desired aggregate statistic.

d::MicroData: A MicroData instance containing agent/individual-level data.

Note

The function handles missing values by replacing them with zeros for the total income variable.

RauhValladaresEsteban2023.policyfunction — Method

policyfunction(ia::Int64, p::ModelParameters)

Returns the policy function for optimal labor supply decisions.

Arguments

ia::Int64: Index of the ability to accumulate human capital.

p::ModelParameters: Instance of ModelParameters containing relevant model parameters.

RauhValladaresEsteban2023.raw2micro — Method

raw2micro(p::ModelParameters; hours::Matrix{Float64}, hk::Matrix{Float64}, empsample::Bool = true)

Converts simulation results of hours worked and human capital to MicroData format optionally applying constraints that allow comparison to NLSY data.

Arguments

p::ModelParameters: Instance of ModelParameters containing relevant parameter values.

hours::Matrix{Float64}: Matrix of simulated hours worked.

hk::Matrix{Float64}: Matrix of simulated human capital.

empsample::Bool = true: (Optional) Flag indicating whether to use treshold to allocate employment indicator.

RauhValladaresEsteban2023.simulate — Function

simulate(ia::Int64, iκs::Matrix{Int64}, n::Vector{Matrix{Float64}}, p::ModelParameters, empsample::Bool = true)

Returns MicroData instance after simulating economy based on the realisations of the staying-at-home shock (iκs) and the policy function (n) for one level of ability to accumulate human capital.

Arguments

ia::Int64: Index of the ability to accumulate human capital.

iκs::Matrix{Int64}: Matrix of the indices of realisations of the staying-at-home shock.

n::Vector{Matrix{Float64}}: Policy function for hours worked.

p::ModelParameters: Instance of ModelParameters containing relevant parameter values.

empsample::Bool = true: Flag indicating whether to use treshold to allocate employment indicator.

RauhValladaresEsteban2023.simulateindices — Method

simulateindices(MC::MarkovChain, nagents::Int64, nperiods::Int64; t₀::Int64 = 100)::Matrix{Int64}

Simulates indices for a given number of agents and periods using a Markov chain.

Arguments

MC::MarkovChain: MarkovChain instance.

nagents::Int64: Number of agents.

nperiods::Int64: Number of periods to simulate.

t₀::Int64 (optional): Amount of periods ignored when returning simulation (default: 100).

Note

The final matrix of simulated indices is returned, starting from period t₀ + 1.

RauhValladaresEsteban2023.simulateraw — Method

simulateraw(ia::Int64, iκs::Matrix{Int64}, n::Vector{Matrix{Float64}}, p::ModelParameters)

Simulates hours worked and human capital based on realisations of the staying-at-home shock (iκs) and the policy function (n) for one level of ability to accumulate human capital.

Arguments

ia::Int64: Index of the ability to accumulate human capital.

iκs::Matrix{Int64}: Matrix of the indices of realisations of the staying-at-home shock.

n::Vector{Matrix{Float64}}: Policy function for hours worked.

p::ModelParameters: Instance of ModelParameters containing model relevant parameter values.

RauhValladaresEsteban2023.statsf — Method

statsf(f, y::Matrix, w::Matrix)

Applies the statistical function f to a matrix y weighted by corresponding elements in the matrix w, after skipping missing values.

Arguments

f: A statistical function to apply to the matrix y weighted by w.

y::Matrix: A matrix containing the data for which the statistic will be calculated.

w::Matrix: A matrix containing the weights to be applied to each element of y. The size of w should match the size of y.

RauhValladaresEsteban2023.statsf — Method

statsf(f, y::Matrix)

Applies the statistical function f to the matrix y, skipping missing values.

Arguments

f: A statistical function to apply to the matrix y.

y::Matrix: A matrix containing the data for which the statistic will be calculated.

RauhValladaresEsteban2023.AgeData — Type

AgeData

A composite type that represents a summary statistic by age for hourly wages, hours worked, and employment status. For example, average wages by age.

Fields

wage::Vector{Float64}: Hourly wage.

hours::Vector{Float64}: Hours worked.

employed::Vector{Float64}: Employment status.

RauhValladaresEsteban2023.MicroData — Type

MicroData

A composite type that represents individual-level data on wages, hours worked, employment status, and individual population weights. Each row corresponds to an agent/individual, and each column represents a period.

Fields

wage::Matrix{Union{Missing, Float64}}: Hourly wage.

hours::Matrix{Union{Missing, Float64}}: Hours worked.

employed::Matrix{Union{Missing, Bool}}: Employment indicator.

weights::Matrix{Union{Missing, Int64}}: Individual frequency weights.

RauhValladaresEsteban2023.ModelParameters — Type

ModelParameters

Composite type that contains all the parameters that describe a model economy.

Fields

Prices

r::Float64: Interest rate.

ω::Float64: Wage rate.

Human capital function parameters

ϕ::Float64: Curvature with respect to hours worked.

δ::Float64: Depreciation.

Preferences

β::Float64: Discount rate.

ψ::Float64: Weight of leisure in utility function.

γ::Float64: Curvature of leisure in utility function.

κ₀::Float64: Intercept value staying at home.

κ₁::Float64: Slope value staying at home.

κ₂::Float64: Curvature value staying at home.

η::Float64: Curvature of human capital in value staying at home.

ρ_κ::Float64: Persistance value staying at home.

σ_κ::Float64: Standard deviation value staying at home.

Model size

gp_h::Int64: Grid points human capital policy funtion.

gp_κ::Int64. Grid points value staying at home.

J::Int64: Number of periods.

nagents::Int64: Number of agents in Monte Carlo per ability level.

nlevels::Int64: umber of ability levels.

Grids & Markov Chains

grid_a::Vector{Float64}: Grid ability.

grid_h₁::Vector{Float64}: Grid initial human capital.

grids_h::Array{Vector{Float64}, 2}: Array of grids of human capital.

grid_κ::Vector{Float64}: Grid stochastic value staying at home.

MCκ::MarkovChain: Markov Chain utility staying at home.

P_κ::Array{Float64,2}: Transition probabilities utility staying at home.

Conversion/comparison to real data

popshare::Vector{Float64}: Population shares; weight of each AFQT decile.

lbhours::Float64: Lower bound hours worked for being employed.

initialage::Int64: Real age at model/data entry.

agerange::UnitRange{Int64}: Range of ages in model/data.