Public Documentation

Index

RauhValladaresEsteban2023.AggregateData

RauhValladaresEsteban2023.catmicrodata

RauhValladaresEsteban2023.catmicrodata

RauhValladaresEsteban2023.gap

RauhValladaresEsteban2023.gap

RauhValladaresEsteban2023.generateparameters

RauhValladaresEsteban2023.groupingmean

RauhValladaresEsteban2023.groupingmean

RauhValladaresEsteban2023.latexcf

RauhValladaresEsteban2023.loadnlsydata

RauhValladaresEsteban2023.meangap

RauhValladaresEsteban2023.modifyparameters

RauhValladaresEsteban2023.modifyparameters

RauhValladaresEsteban2023.plotmvsd

RauhValladaresEsteban2023.plotmvsd

RauhValladaresEsteban2023.simulatewages

RauhValladaresEsteban2023.simulatewages

RauhValladaresEsteban2023.solveconstantls

RauhValladaresEsteban2023.solvemodel

RauhValladaresEsteban2023.statsfbyage

RauhValladaresEsteban2023.statsfbyage

RauhValladaresEsteban2023.statsfbyage

RauhValladaresEsteban2023.sterror

Public Interface

RauhValladaresEsteban2023.catmicrodata — Method

catmicrodata(grps::Vector, v::Vector{MicroData})

Concatenates multiple MicroData instances based on groupings provided in grps.

Arguments

grps::Vector: Vector of groupings, where each element is a single index (an Int64) or a range of indices (UnitRange{Int64}). The indices correspond to the positions of MicroData instances in the input vector v.

v::Vector{MicroData}: Vector of MicroData instances to be concatenated.

Notes

The length of the grps vector must be less than or equal to the length of the v vector.

Each element in grps specifies a grouping of MicroData instances to be concatenated together.

The output vector z has the same length as the grps vector.

RauhValladaresEsteban2023.catmicrodata — Method

catmicrodata(v::Vector{MicroData})

Concatenates multiple instances of MicroData structures stored in vector (v) and returns a new MicroData instance with the concatenated data.

Arguments

v::Vector{MicroData}: A vector of MicroData instances to be concatenated.

Notes

The function assumes that all MicroData instances in the input vector have the same structure.

The concatenation is performed by vertically stacking the corresponding fields of each MicroData instance.

The resulting MicroData instance will have the same field names as the input instances.

RauhValladaresEsteban2023.gap — Method

gap(b::AggregateData, w::AggregateData)

Calculates the gap between two AggregateData instances b and w for each corresponding field.

Arguments

b::AggregateData: The numerator AggregateData instance.

w::AggregateData: The denominator AggregateData instance.

RauhValladaresEsteban2023.gap — Method

gap(b::Float64, w::Float64)

Calculates the gap between the two values b and w as $1 - \frac{b}{w}$.

Arguments

b::Float64: The numerator value.

w::Float64: The denominator value.

RauhValladaresEsteban2023.generateparameters — Method

generateparameters(group::String)

Generates an instance of ModelParameters by loading and processing input data specific to the provided group.

Arguments

group::String: The name of the predefined group. Valid options are 'Black' or 'White'.

RauhValladaresEsteban2023.groupingmean — Method

groupingmean(grps::Vector, v::Vector{AgeData}, p::ModelParameters)

Computes the weighted mean of each field in the vector of AgeData instances v for specific groupings defined by the indices in grps, using the popshare field from the ModelParameters object p as weights.

Arguments

grps::Vector: Vector of groupings, where each element is a single index (an Int64) or a range of indices (UnitRange{Int64}). The indices correspond to the positions of MicroData instances in the input vector v.

v::Vector{AgeData}: A vector of AgeData instances.

p::ModelParameters: A ModelParameters instance.

Notes

The length of the grps vector must be less than or equal to the length of the v vector.

Each element in grps specifies a grouping of MicroData instances to be concatenated together.

The output vector z has the same length as the grps vector.

RauhValladaresEsteban2023.groupingmean — Method

groupingmean(m::Vector{AgeData}, w::Vector{Float64})

Computes the weighted mean of each field in the vector of AgeData instnaces m using the corresponding weights in the vector w.

Arguments

m::Vector{AgeData}: A vector of AgeData instances.

w::Vector{Float64}: A vector of weights. The length of w must be equal to the length of m.

RauhValladaresEsteban2023.latexcf — Method

latexcf(gaps::Dict{String, AggregateData})

Generates a LaTeX table with counterfactual results based on the provided gaps dictionary.

Arguments

gaps::Dict{String, AggregateData}: Dictionary containing counterfactual results for different scenarios. The keys are scenario names, and the values are AggregateData instances representing the gaps.

Note

This function generates a LaTeX document and saves it as a file named "counterfactuals.tex" in the "tables" directory. The LaTeX document contains a table displaying the counterfactual results.

RauhValladaresEsteban2023.loadnlsydata — Function

loadnlsydata(group::String = "")

Loads NLSY (National Longitudinal Survey of Youth) data from the nlsy_data folder and returns it as an vector of MicroData.

Arguments

group::String: (optional) The group to which the data should be restricted. Valid values are "Black", "White", or an empty string (default).

Notes

The function reads all files in the nlsy_data folder and assumes that each file contains AFQT decile data in the expected format.

If group is specified, the data will be restricted to the specified group ("black" or "white") for each file.

Examples

data = loadnlsydata()  # Load NLSY data for all groups
data = loadnlsydata("black")  # Load NLSY data for the black group only

RauhValladaresEsteban2023.meangap — Method

meangap(b::Vector{MicroData}, wb::Vector{Float64}, w::Vector{MicroData}, ww::Vector{Float64})

Calculates the mean gap between the two vectors of MicroData instances, b and w, weighted by the corresponding weight vectors wb and ww.

Arguments

b::Vector{MicroData}: The numerator vector of MicroData instances.

wb::Vector{Float64}: The weight vector corresponding to the numerator vector.

w::Vector{MicroData}: The denominator vector of MicroData instances.

ww::Vector{Float64}: The weight vector corresponding to the denominator vector.

RauhValladaresEsteban2023.modifyparameters — Method

modifyparameters(p::ModelParameters; np...)

Creates a new instance of ModelParameters by modifying the specified fields while copying the remaining fields from the input p.

Arguments:

p::ModelParameters: The original instance of ModelParameters.

np...: Keyword arguments representing the fields of ModelParameters to be modified along with their desired values. Example: ψ = 1.0, γ = 2.0.

Note

The function assumes that the specified keyword arguments correspond to the fields of ModelParameters.

RauhValladaresEsteban2023.modifyparameters — Method

modifyparameters(group::String; np...)

Creates a new instance of ModelParameters by modifying the fields specified in np... while copying the remaining fields from the generateparameters(group) function.

Arguments

group::String: The name of the predefined group. It can be either "Black" or "White".

np...: Keyword arguments representing the fields of ModelParameters with their desired values. Example: ψ = 1.0, γ = 2.0.

Note

The function assumes that the specified keyword arguments correspond to the fields of ModelParameters.

RauhValladaresEsteban2023.plotmvsd — Method

plotmvsd(
    fn::Symbol,
    m::Vector{AgeData},
    d::Vector{AgeData};
    de::Union{Vector{AgeData}, Nothing} = nothing,
    plt_ttl::String = "",
    lyt = (1, 5),
    sz = (2000, 500),
    mrgn = (50, :px),
    ylms = (0, NaN),
    haxis = 1:length(d),
    spttls = fill("", length(d)),
    mlbl = "Model",
    dlbl = "Data",
    xlbl = "Age",
    flllph = 0.1,
    zα = 1.96
    )

Plots multiple subplots the model simulated data and real data with error bars representing the standard error.

Arguments

fn::Symbol: Field name to plot from AgeData instances.

m::Vector{AgeData}: Vector of AgeData structures of model simulated data.

d::Vector{AgeData}: Vector of AgeData structures of real data.

de::Union{Vector{AgeData}, Nothing} = nothing: Optional vector of AgeData instances representing standard errors for the data. If not provided (default), error bars will not be shown.

plt_ttl::String = "": Optional title for the overall plot.

lyt::Tuple{Int64, Int64} = (1, 5): Layout of subplots in rows and columns.

sz::Tuple{Int64, Int64} = (2000, 500): Size of the overall plot.

mrgn::Tuple{Int64, Symbol} = (50, :px): Margin size and unit for the overall plot.

ylms::Tuple{Float64, Float64} = (0, NaN): Y-axis limits for the overall plot.

haxis: Horizontal axis values for each subplot. Defaults to 1:length(d).

spttls::Vector{String} = fill("", length(d)): Titles for each subplot. Defaults to an empty string for each subplot.

mlbl::String = "Model": Label for the model predictions in the legend.

dlbl::String = "Data": Label for the data in the legend.

xlbl::String = "Age": Label for the x-axis.

flllph::Float64 = 0.1: Fill alpha value for the error bars. Defaults to 0.1.

zα::Float64 = 1.96: Z-value corresponding to the desired level of confidence for the error bars. Defaults to 1.96, representing the 95% confidence level.

RauhValladaresEsteban2023.plotmvsd — Method

plotmvsd(
    m::Vector{Float64},
    d::Vector{Float64};
    de::Union{Vector{Float64}, Nothing} = nothing,
    haxis = 1:length(d),
    ttl = "",
    mlbl = "Model",
    dlbl = "Data",
    xlbl = "Age",
    flllph = 0.1,
    zα = 1.96
    )

Plots the model simulated data and real data with error bars representing the standard error.

Arguments

m::Vector{Float64}: Model simulated data.

d::Vector{Float64}: Real data.

de::Union{Vector{Float64}, Nothing} = nothing: Optional vector of standard errors for the data. If not provided (default), error bars will not be shown.

haxis: Horizontal axis values for the plot. Defaults to 1:length(d).

ttl::String = "": Optional title for the plot.

mlbl::String = "Model": Label for the model predictions in the legend.

dlbl::String = "Data": Label for the data in the legend.

xlbl::String = "Age": Label for the x-axis.

flllph::Float64 = 0.1: Fill alpha value for the error bars. Defaults to 0.1.

zα::Float64 = 1.96: Z-value corresponding to the desired level of confidence for the error bars. Defaults to 1.96, representing the 95% confidence level.

RauhValladaresEsteban2023.simulatewages — Method

simulatewages(hours::Matrix, a::Float64, h₁::Float64, p::ModelParameters)

Simulates wages based on hours worked, the ability to accumulate human capital, and the initial level of human capital.

Arguments

hours::Matrix: Matrix containing hours worked by agents (rows) for each period (columns).

a::Float64: Ability to accumulate human capital.

h₁::Float64: Initial level of human capital.

p::ModelParameters: Instance of ModelParameters containing relevant model parameters.

RauhValladaresEsteban2023.simulatewages — Method

simulatewages(v::Vector{MicroData}, p::ModelParameters)

Simulates wages for each level of ability and initial human capital.

Arguments

v::Vector{MicroData}: Vector of MicroData containing amount of hours worked.

p::ModelParameters: Instance of ModelParameters containing relevant model parameters.

RauhValladaresEsteban2023.solveconstantls — Method

solveconstantls(pb::ModelParameters, pw::ModelParameters)

Solves the model using the labor supply decisions of the economy defined by pb while using the human capital parameters of pw.

Arguments

pb::ModelParameters: Instance of ModelParameters to compute labor supply decisions.

pw::ModelParameters: Instance of ModelParameters to use human capital parameters.

Note

This function returns a vector of simulated data for each level of ability/h₁. Each element in the vector represents the simulated data for one level and is of type MicroData.

RauhValladaresEsteban2023.solvemodel — Function

solvemodel(p::ModelParameters, empsample::Bool = true)

Solves the model based on the parameters provided.

Arguments

p::ModelParameters: Instance of ModelParameters containing the model's parameter values.

empsample::Bool = true: (Optional) Flag indicating whether to use treshold to allocate employment indicator.

Note

This function returns a vector of simulated data for each level of ability/h₁. Each element in the vector represents the simulated data for one level and is of type MicroData.

RauhValladaresEsteban2023.statsfbyage — Method

statsfbyage(f, y::Matrix, w::Matrix)

Applies the statistical function f to each column of the matrix y weighted by corresponding elements in the matrix w, skipping missing values.

Arguments

f: A statistical function to apply to each column of the matrix y weighted by w.

y::Matrix: A matrix containing the data for which the statistic will be calculated.

w::Matrix: A matrix containing the weights to be applied to each element of y. The size of w should match the size of y.

RauhValladaresEsteban2023.statsfbyage — Method

statsfbyage(f, y::Matrix)

Applies the statistical function f to each column of the matrix y, treating missing values as skipped.

Arguments

f: A statistical function to apply to each column of y.

y::Matrix: A matrix containing the data for which the statistics will be calculated.

RauhValladaresEsteban2023.statsfbyage — Method

statsfbyage(f, d::MicroData)

Applies the statistical function f to each column of the wage, hours, and employed fields in the MicroData instance d weighted by the corresponding elements in the weights field, after skipping missing values.

Arguments

f: A statistical function to apply to each column of the fields in d.

d::MicroData: A MicroData object containing wage, hours, employed, and weights fields.

RauhValladaresEsteban2023.sterror — Method

sterror(v::AbstractVector{<:Real}, w::AbstractWeights)

Calculates the standard error of a weighted sample vector v using the provided weights w.

Arguments

v::AbstractVector{<:Real}: A vector of real values representing the sample data.

w::AbstractWeights: An abstract type representing the weights associated with each data point in v.

Notes

The weights w must have the same length as the vector v.

The standard error is calculated as the standard deviation of v weighted by w, divided by the square root of the sample size (length(w)).

RauhValladaresEsteban2023.AggregateData — Type

AggregateData

Composite type that contains a summary statistic for hourly wages among the employed, hours worked among the employed, employment rate, income among all, and income among the employed. For example, average wages.

Fields

wage::Float64: Average wage among the employed.

hours::Float64: Hours worked among the employed.

employment::Float64: Employment rate.

incomeall::Float64: Average income among all.

incomeemp::Float64: Average income among the employed.